Refactored driver to unify data types inside driver module
It became too hard to shift data around. For example, I need to print
expected statements, which rely on the metadata for k3_dist, but it just
wasn't available at a late stage. I unified all the different data types
that the driver moves around into one type, and this also makes the
tests conform to the structure of the other execution types.