Changed the way corrective and do_complete is done
We now fully separate the delta computation from the save to the lmap.
The save part is common to both do_complete and correctives: we need to
compute the full value (with initialization if needed) for the current vid,
and to propagate the delta value to all future vids for the same key.

In a more dynamic language (such as the new K3), we can reduce greatly
the number of global functions by using a reference to the specific map
to be updated.