Changed master log to use statements rather than triggers
Also moved the log update from rcv_fetch to send_push, even though this
forces us to use statements rather than triggers. This is important for
the log-update-message-send action to be atomic ie. we don't invoke
other triggers between these actions, so we can't send another push in
the middle.