shuffle: remove bound vars from rmap-lmap binding pattern
As far as shuffle is concerned, bound vars are irrelevant for binding
patterns between lmap and rmap. They might as well be disconnected. We
only care about loops, which only occur with free vars.

This isn't the case for map buffer optimization, which cares about any
connection in sharding patterns.