attempt to fix queries 5,7,8,10
The occasional errors seem to result from an ordering that causes
statements with multiple loop vars to break, due to the fact that we
send information conservatively (ie, some rhs maps are sharded to the
lhs map coordinates where they don't belong). We add an lhs test before
writing the data and after calculation. An earlier filter would be
better.