dist: push thinning: make messages a stmt with a count
There's no reason to send metadata about the exact stmt/map combos.
That's specified in unique_poly (ie. data). All we need is a counter,
and the counter values are small enough we can encode them in the tag.
This work is also useful for batching.
TODO: make no-corrective pushes barrier on the send for the whole
statement, so we don't break up stmts into maps.